dynamic programming?
My mission was to create some fns to to enable me to start and stop the game from the repl. Right now when the window is closed, the whole jvm shuts down, and that is not what is helpful when working with the repl. But somehow this does not work at all. Stopping throws an exception :(